Time: 24 ms (92.62%) | Memory: 132.1 MB (80.07%) - LeetSync

+++ 3031-construct-product-matrix/construct-product-matrix.cpp
@@ -7,7 +7,7 @@ public:
         int m = grid[0].size();
 
         vector<vector<int>>suff(n , vector<int>(m));
-        long long suffix = 1;
+        long  suffix = 1;
         for(int i = n-1 ; i >= 0 ; i--){
             for(int j = m-1 ; j >= 0 ; j--){    
                 suff[i][j] = suffix;
